Removed 2MB homebrew limit

+++ types.h
@@ -22,7 +22,7 @@ typedef long long int s64_t;        ///< Signed 64-bit type
  *  @{
  */
 typedef int PsvUID;            ///< PsvUID type
-typedef long int PsvOff;       ///< PsvOff type
+typedef long long int PsvOff;       ///< PsvOff type
 typedef int PsvSSize;          ///< PsvSSize type
 /** @}*/
